:truck: Move `get_scheduler` out of tags catchall header
