+++ include/async/read_env.hpp
@@ -3,6 +3,7 @@
 #include <async/completes_synchronously.hpp>
 #include <async/concepts.hpp>
 #include <async/env.hpp>
+#include <async/get_scheduler.hpp>
 #include <async/stop_token.hpp>
 #include <async/tags.hpp>
 #include <async/type_traits.hpp>
