+++ include/async/tags.hpp
@@ -85,26 +85,4 @@ constexpr inline struct start_t {
         return {};
     }
 } start{};
-
-struct get_scheduler_t : forwarding_query_t {
-    template <typename T>
-        requires true
-    constexpr auto operator()(T &&t) const noexcept(
-        noexcept(std::forward<T>(t).query(std::declval<get_scheduler_t>())))
-        -> decltype(std::forward<T>(t).query(*this)) {
-        return std::forward<T>(t).query(*this);
-    }
-
-    template <typename T>
-    constexpr auto operator()(T &&) const -> failure_t<T> {
-        static_assert(stdx::always_false_v<T>,
-                      "No function call for get_scheduler: does the argument "
-                      "provide a get_scheduler_t query?");
-        return {};
-    }
-};
-
-template <typename E> auto get_scheduler(E &&e) {
-    return get_scheduler_t{}(std::forward<E>(e));
-}
 } // namespace async
